+++ proto/src/prost/tendermint.store.rs
@@ -1,0 +1,7 @@
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct BlockStoreState {
+    #[prost(int64, tag="1")]
+    pub base: i64,
+    #[prost(int64, tag="2")]
+    pub height: i64,
+}