+++ testgen/src/consensus.rs
@@ -1,6 +1,7 @@
 use tendermint::{block, consensus, evidence, public_key::Algorithm};
 
-/// Default consensus params modeled after Go code; but it's not clear how to go to a valid hash from here
+/// Default consensus params modeled after Go code; but it's not clear how to go to a valid hash
+/// from here
 pub fn default_consensus_params() -> consensus::Params {
     consensus::Params {
         block: block::Size {